shape.java
1. Develop an abstract class called Shape having two data members base and height of type
double. Derive three specific sub-classes Triangle, Circle and Rectangle from the base Shape. Add
to the base class Shape, a relevant constructor to initialize base class data members and a method
computeArea() to compute and display the area of figures. Make computeArea() as an abstract
method and define this method in the sub-classes to compute area specific to each of the subclasses. (Area of Triangle=0.5*base*height, Area of Rectangle=base*height and Area of
Circle=pi*radius*radius).